+++ src/Authorization/DebugCommand.php
@@ -31,29 +31,37 @@ class DebugCommand extends Command implements LoggerAwareInterface
     {
         $this->setName('dbp:relay:core:auth-debug');
         $this->setDescription('Shows various information about the authorization providers');
-        $this->addArgument('username', InputArgument::OPTIONAL, 'username');
+        $this->addArgument('attribute', InputArgument::REQUIRED, 'the user attribute name to query');
+        $this->addArgument('user_id', InputArgument::OPTIONAL, 'the user identifier to query');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $username = $input->getArgument('username');
+        try {
+            $attribute = $input->getArgument('attribute');
+            $userId = $input->getArgument('user_id');
 
-        $attrs = $this->userAttributeMuxer->getAvailableAttributes();
-        $all = [];
-        $default = new \stdClass();
-        sort($attrs, SORT_STRING | SORT_FLAG_CASE);
-        foreach ($attrs as $attr) {
-            $all[$attr] = $this->userAttributeMuxer->getAttribute($username, $attr, $default);
-        }
-
-        // Now print them out
-        $output->writeln('<fg=blue;options=bold>[Authorization attributes]</>');
-        foreach ($all as $attr => $value) {
-            if ($value === $default) {
-                $output->writeln('<fg=green;options=bold>'.$attr.'</> = <fg=magenta;options=bold>\<N/A\></>');
-            } else {
-                $output->writeln('<fg=green;options=bold>'.$attr.'</> = '.json_encode($value));
+            $attrs = [$attribute];
+            $all = [];
+            $default = new \stdClass();
+            sort($attrs, SORT_STRING | SORT_FLAG_CASE);
+            foreach ($attrs as $attr) {
+                $all[$attr] = $this->userAttributeMuxer->getAttribute($userId, $attr, $default);
             }
+
+            // Now print them out
+            $output->writeln('<fg=blue;options=bold>[Authorization attributes]</>');
+            foreach ($all as $attr => $value) {
+                if ($value === $default) {
+                    $output->writeln('<fg=green;options=bold>'.$attr.'</> = <fg=magenta;options=bold>\<N/A\></>');
+                } else {
+                    $output->writeln('<fg=green;options=bold>'.$attr.'</> = '.json_encode($value));
+                }
+            }
+        } catch (\Throwable $e) {
+            $output->writeln('<fg=red;options=bold>An error occurred: '.$e->getMessage().'</>');
+
+            return 1;
         }
 
         return 0;